requirements.txt
flask==3.0.3 
numpy==1.26.4 
pillow==10.3.0 
tensorflow-cpu==2.16.1 
gunicorn==22.0.0 
requests==2.31.0 
